Fixing valid address check, wallet existence, subscripton own file.
